requirements.txt
anyio==4.12.0
asgiref==3.10.0
CacheControl==0.14.4
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.2.7
firebase_admin==7.1.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
msgpack==1.1.2
proto-plus==1.26.1
protobuf==6.33.1
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.5
rsa==4.9.1
sqlparse==0.5.3
Telethon==1.42.0
typing_extensions==4.15.0
urllib3==2.5.0
vk_api==11.10.0